+++ src/models/Recipe.php
@@ -13,10 +13,14 @@ namespace nystudio107\recipe\models;
 
 use nystudio107\recipe\helpers\Json;
 use nystudio107\recipe\helpers\PluginTemplate;
+use nystudio107\seomatic\Seomatic;
+use nystudio107\seomatic\models\MetaJsonLd;
 
 use Craft;
 use craft\base\Model;
+use craft\helpers\StringHelper;
 use craft\helpers\Template;
+use craft\validators\ArrayValidator;
 
 use Twig\Markup;
 
@@ -29,6 +33,9 @@ class Recipe extends Model
 {
     // Constants
     // =========================================================================
+
+    const SEOMATIC_PLUGIN_HANDLE = 'seomatic';
+    const MAIN_ENTITY_KEY = 'mainEntityOfPage';
 
     const US_RDA = [
         'calories' => 2000,
@@ -73,6 +80,16 @@ class Recipe extends Model
     /**
      * @var string
      */
+    public $recipeCategory;
+
+    /**
+     * @var string
+     */
+    public $recipeCuisine;
+
+    /**
+     * @var string
+     */
     public $skill = 'intermediate';
 
     /**
@@ -94,6 +111,11 @@ class Recipe extends Model
      * @var array
      */
     public $directions = [];
+
+    /**
+     * @var array
+     */
+    public $equipment = [];
 
     /**
      * @var int
@@ -209,6 +231,8 @@ class Recipe extends Model
             ['name', 'string'],
             ['name', 'default', 'value' => ''],
             ['description', 'string'],
+            ['recipeCategory', 'string'],
+            ['recipeCuisine', 'string'],
             ['skill', 'string'],
             ['serves', 'integer'],
             ['imageId', 'integer'],
@@ -227,17 +251,24 @@ class Recipe extends Model
             ['sugarContent', 'integer'],
             ['transFatContent', 'integer'],
             ['unsaturatedFatContent', 'integer'],
+            [
+                [
+                    'ingredients',
+                    'directions',
+                    'equipment',
+                ],
+                ArrayValidator::class,
+            ],
+
         ];
     }
 
     /**
-     * Render the JSON-LD Structured Data for this recipe
+     * Return the JSON-LD Structured Data for this recipe
      *
-     * @param bool $raw
-     *
-     * @return string|\Twig_Markup
+     * @return array
      */
-    public function renderRecipeJSONLD($raw = true)
+    public function getRecipeJSONLD(): array
     {
         $recipeJSONLD = [
             'context' => 'http://schema.org',
@@ -245,9 +276,12 @@ class Recipe extends Model
             'name' => $this->name,
             'image' => $this->getImageUrl(),
             'description' => $this->description,
+            'recipeCategory' => $this->recipeCategory,
+            'recipeCuisine' => $this->recipeCuisine,
             'recipeYield' => $this->getServes(),
             'recipeIngredient' => $this->getIngredients('imperial', 0, false),
             'recipeInstructions' => $this->getDirections(false),
+            'tool' => $this->getEquipment(false),
         ];
         $recipeJSONLD = array_filter($recipeJSONLD);
 
@@ -313,7 +347,62 @@ class Recipe extends Model
             $recipeJSONLD['totalTime'] = 'PT' . $this->totalTime . 'M';
         }
 
-        return $this->renderJsonLd($recipeJSONLD, $raw);
+        return $recipeJSONLD;
+    }
+
+    /**
+     * Create the SEOmatic MetaJsonLd object for this recipe
+     *
+     * @param bool $add
+     * @param null $key
+     * @return null|MetaJsonLd
+     */
+    public function createRecipeMetaJsonLd($key = null, bool $add = true)
+    {
+        $result = null;
+        if (Craft::$app->getPlugins()->getPlugin(self::SEOMATIC_PLUGIN_HANDLE)) {
+            $seomatic = Seomatic::getInstance();
+            if ($seomatic !== null) {
+                $recipeJson = $this->getRecipeJSONLD();
+                // If we're adding the MetaJsonLd to the container, and no key is provided, give it a random key
+                if ($add && $key === null) {
+                    try {
+                        $key = StringHelper::UUID();
+                    } catch (\Exception $e) {
+                        // That's okay
+                    }
+                }
+                if ($key !== null) {
+                    $recipeJson['key'] = $key;
+                }
+                // If the key is `mainEntityOfPage` add in the URL
+                if ($key === self::MAIN_ENTITY_KEY) {
+                    $mainEntity = Seomatic::$plugin->jsonLd->get(self::MAIN_ENTITY_KEY);
+                    if ($mainEntity) {
+                        $recipeJson[self::MAIN_ENTITY_KEY] = $mainEntity[self::MAIN_ENTITY_KEY];
+                    }
+                }
+
+                $result = Seomatic::$plugin->jsonLd->create(
+                    $recipeJson,
+                    $add
+                );
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Render the JSON-LD Structured Data for this recipe
+     *
+     * @param bool $raw
+     *
+     * @return string|\Twig_Markup
+     */
+    public function renderRecipeJSONLD($raw = true)
+    {
+        return $this->renderJsonLd($this->getRecipeJSONLD(), $raw);
     }
 
     /**
@@ -548,6 +637,29 @@ class Recipe extends Model
                     $direction = Template::raw($direction);
                 }
                 $result[] = $direction;
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Get all of the equipment for this recipe
+     *
+     * @param bool $raw
+     *
+     * @return array
+     */
+    public function getEquipment($raw = true)
+    {
+        $result = [];
+        if (!empty($this->equipment)) {
+            foreach ($this->equipment as $row) {
+                $equipment = $row['equipment'];
+                if ($raw) {
+                    $equipment = Template::raw(equipment);
+                }
+                $result[] = $equipment;
             }
         }