+++ src/translations/en/recipe.php
@@ -19,5 +19,15 @@ return [
     'Review' => 'Review',
     '{name} plugin loaded' => '{name} plugin loaded',
     'Error rendering template string -> {error}' => 'Error rendering template string -> {error}',
-    'Error rendering `{template}` -> {error}' => 'Error rendering `{template}` -> {error}'
+    'Error rendering `{template}` -> {error}' => 'Error rendering `{template}` -> {error}',
+    'API Application Key' => 'API Application Key',
+    'Failed to generate nutritional information for {count} entries.' => 'Failed to generate nutritional information for {count} entries.',
+    'API credentials do not exist in plugin settings.' => 'API credentials do not exist in plugin settings.',
+    'An application ID for the Edamam Nutrition Analysis API.' => 'An application ID for the Edamam Nutrition Analysis API.',
+    'No entries found in the section with handle `{handle}`.' => 'No entries found in the section with handle `{handle}`.',
+    'A section handle must be provided using --section.' => 'A section handle must be provided using --section.',
+    'Generating nutritional information for {count} entries...' => 'Generating nutritional information for {count} entries...',
+    'The [Edamam Nutrition Analysis API]({url}) can be used to fetch nutritional information from your ingredients.' => 'The [Edamam Nutrition Analysis API]({url}) can be used to fetch nutritional information from your ingredients.',
+    'A field handle must be provided using --field.' => 'A field handle must be provided using --field.',
+    'Successfully generated nutritional information for {count} entries.' => 'Successfully generated nutritional information for {count} entries.'
 ];